+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/Calculations/DatabricksCalculationParametersFactory.cs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Batches.Application.Model.Batches;
-using Microsoft.Azure.Databricks.Client;
 using Microsoft.Azure.Databricks.Client.Models;
 
 namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.Calculations;