+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.IntegrationTests/Infrastructure/CalculationResults/WholesaleResultQueriesTests.cs
@@ -21,11 +21,9 @@ using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationR
 using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SqlStatements.DeltaTableConstants;
 using Energinet.DataHub.Wholesale.CalculationResults.IntegrationTests.Fixtures;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.WholesaleResults;
-using Energinet.DataHub.Wholesale.Common.Databricks.Options;
 using FluentAssertions;
 using FluentAssertions.Execution;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using Moq;
 using Xunit;
 
@@ -79,12 +77,12 @@ public class WholesaleResultQueriesTests : IClassFixture<DatabricksSqlStatementA
         using var assertionScope = new AssertionScope();
         var actualHourlyAmount = actual.Single(row => row.Id.ToString() == HourlyTariffCalculationResultId);
         actualHourlyAmount.AmountType.Should().Be(AmountType.AmountPerCharge);
-        actualHourlyAmount.ChargeResolution.Should().Be(ChargeResolution.Hour);
+        actualHourlyAmount.Resolution.Should().Be(Resolution.Hour);
         actualHourlyAmount.TimeSeriesPoints.First().Amount.Should().Be(decimal.Parse(DefaultHourlyAmount, CultureInfo.InvariantCulture));
 
         var actualMonthlyAmount = actual.Single(row => row.Id.ToString() == MonthlyAmountTariffCalculationResultId);
         actualMonthlyAmount.AmountType.Should().Be(AmountType.MonthlyAmountPerCharge);
-        actualMonthlyAmount.ChargeResolution.Should().Be(ChargeResolution.Month);
+        actualMonthlyAmount.Resolution.Should().Be(Resolution.Month);
         actualMonthlyAmount.TimeSeriesPoints.First().Amount.Should().Be(decimal.Parse(DefaultMonthlyAmount, CultureInfo.InvariantCulture));
     }
 
@@ -94,7 +92,7 @@ public class WholesaleResultQueriesTests : IClassFixture<DatabricksSqlStatementA
             calculationId: CalculationId,
             calculationResultId: HourlyTariffCalculationResultId,
             chargeType: "tariff",
-            chargeResolution: "PT1H",
+            resolution: "PT1H",
             amountType: "amount_per_charge",
             meteringPointType: "production",
             settlementMethod: null,
@@ -104,7 +102,7 @@ public class WholesaleResultQueriesTests : IClassFixture<DatabricksSqlStatementA
             calculationId: CalculationId,
             calculationResultId: MonthlyAmountTariffCalculationResultId,
             chargeType: "tariff",
-            chargeResolution: "P1M",
+            resolution: "P1M",
             amountType: "monthly_amount_per_charge",
             meteringPointType: null,
             settlementMethod: null,