+++ source/dotnet/wholesale-integration-events-package/Contracts.Documentation/release-notes/release-notes.md
@@ -1,5 +1,17 @@
 # Wholesale Contracts Release notes
 
+## Version 5.4.0
+
+- Added `EnergyResultProducedV2` with `EventMinorVersion=0`. We also deleted `EnergyResultProducedV1`, but as it was never used by anyone we only bump our minor version of the package.
+
+## Version 5.3.0
+
+- Updated `AmountPerChargeResultProducedV1`: `quantity` is not nullable. Bumped `EventMinorVersion` to `3`.
+
+## Version 5.2.0
+
+- Bump versions of NuGet package dependencies.
+
 ## Version 5.1.0
 
 Added 'Currency' to 'AmountPerChargeResultProducedV1' and 'MonthlyAmountPerChargeResultProducedV1